Commit - 202301231603
Added project files from assignment 6 and changed document signing.

Updated the README.md file to reflect the project changes.

Updated gradle to 7.4 from 7.3.3

(Activity_inbox_fragment.xml) - Added the inbox fragment activity with a recyclerview to hold the fragment containing the inbox list

(activity_list_multi_selection.xml) - Switched from the recyclerView to a constraint layout for the fragment placeholders

(AdapterListBasic.java) - Added a updateList function to call the notifyDataSetChanged method.

(forward_email_fragment.xml) - Added the email fragment for forwarding emails.

(forward_fragment.xml) - Added the forward button in the form of an activity to the inbox list.

(ForwardEmailFragment.java) - Added the ForwardEmailFragment class object to handle creating a new instance of the forwarded email fragment and load the data from the shared data model.

(ForwardFragment.java) - Added the ForwardFragment class to inflate the forwarded email button activity.

(Inbox.java) - Converted the Inbox class object to Parcelable that way it can be passed via newInstance and onSendBtnClicked in and out of the fragment

(InboxListFragment.java) - Moved the items from the previous example/assignment to the InboxListFragment class. This moved the Inbox list to a fragment as compared to in the activity as the last assignment used.

(MainActivity.java) - Moved the Inbox list to a fragment by moving it from the Main Activity to its own class, InboxListFragment.java)

(SharedDataModel) - Added the shared data model to be able to create a pipeline between the different fragments used in the app. This allows the transfer of data to the forward email fragment from the Inbox list and back.

(toolbar_menu.xml) - Added the toolbar menu item so that way the forward email button would be displayed as per the specifications.

Cleaned | Synced | ReBuilt | Debugged